chore(date-range-input): move to packages
