+++ packages/components/inputs/date-range-input/src/date-range-input.js
@@ -20,11 +20,11 @@ import {
   formatRange,
   parseInputToDate,
 } from '@commercetools-uikit/calendar-utils';
-import CalendarBody from '../../internals/calendar-body';
-import CalendarMenu from '../../internals/calendar-menu';
-import CalendarHeader from '../../internals/calendar-header';
-import CalendarContent from '../../internals/calendar-content';
-import CalendarDay from '../../internals/calendar-day';
+import CalendarBody from '../../../../../src/components/internals/calendar-body';
+import CalendarMenu from '../../../../../src/components/internals/calendar-menu';
+import CalendarHeader from '../../../../../src/components/internals/calendar-header';
+import CalendarContent from '../../../../../src/components/internals/calendar-content';
+import CalendarDay from '../../../../../src/components/internals/calendar-day';
 import messages from './messages';
 
 const parseRangeText = (text, locale) => {
