+++ packages/components/inputs/date-range-input/src/date-range-input.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent } from '../../../../test/test-utils';
+import { render, fireEvent } from '../../../../../test/test-utils';
 import DateRangeInput from './date-range-input';
 
 // This component is used to enable easy testing.
