+++ packages/components/inputs/date-range-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './date-range-input';
+
+export { default as version } from './version';
