--- src/components/inputs/date-range-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './date-range-input';
